# apps/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountScoreBand": {
        "description": "`account_score.band`.",
        "enum": [
          "HOT",
          "WARM",
          "COLD"
        ],
        "title": "AccountScoreBand",
        "type": "string"
      },
      "AccountScoreStanding": {
        "description": "`account_score.standing`.",
        "enum": [
          "RANKED",
          "BELOW_FIT",
          "DISQUALIFIED",
          "CUSTOMER"
        ],
        "title": "AccountScoreStanding",
        "type": "string"
      },
...
      "AlertView": {
        "additionalProperties": false,
        "description": "[`AlertView`](/architecture/interfaces.md#alertview), the response of `API-48` and\n`API-49`.",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/AlertRef"
          },
          "acknowledged_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged At"
          },
          "acknowledged_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acknowledged By Name"
          },
          "band_change": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlertBandChange"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AlertFinding"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/AlertKind"
          },
          "service": {
            "$ref": "#/components/schemas/AlertRef"
          }
        },
        "required": [
          "id",
          "created_at",
          "acknowledged_at",
          "kind",
          "account",
          "service",
          "finding",
          "band_change",
          "acknowledged_by_name"
        ],
        "title": "AlertView",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "The `{\"error\": {...}}` shape of [Conventions](/architecture/interfaces.md#conventions),\nnamed `ErrorEnvelope` in the OpenAPI document.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "FeedbackCreate_FindingFeedbackVerdict_": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "title": "Note",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/FindingFeedbackVerdict"
          }
        },
        "required": [
          "verdict"
        ],
        "title": "FeedbackCreate[FindingFeedbackVerdict]",
        "type": "object"
      },
      "FeedbackCreate_LeadFeedbackVerdict_": {
        "additionalProperties": false,
        "properties": {
          "note": {
            "title": "Note",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/LeadFeedbackVerdict"
          }
        },
        "required": [
          "verdict"
        ],
        "title": "FeedbackCreate[LeadFeedbackVerdict]",
        "type": "object"
      },
...
      "FindingView": {
        "additionalProperties": false,
        "description": "[`FindingView`](/architecture/interfaces.md#findingview), the response of `API-47`. Also\nplan task 13's `API-42` response shape; that task imports this model and must not define a\nsecond one.",
        "properties": {
          "account_id": {
            "format": "uuid",
            "title": "Account Id",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "decided_by": {
            "$ref": "#/components/schemas/FindingDecidedBy"
          },
          "document": {
            "$ref": "#/components/schemas/FindingViewDocument"
          },
          "feedback": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FindingViewFeedback"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "option": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FindingViewOption"
              },
              {
                "type": "null"
              }
            ]
          },
          "points": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Points"
          },
          "question": {
            "$ref": "#/components/schemas/FindingViewQuestion"
          },
          "question_revision": {
            "title": "Question Revision",
            "type": "integer"
          },
          "quote": {
            "title": "Quote",
            "type": "string"
          },
          "quote_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote En"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "service_id": {
            "format": "uuid",
            "title": "Service Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/FindingStatus"
          },
          "strength": {
            "$ref": "#/components/schemas/FindingStrength"
          }
        },
        "required": [
          "id",
          "account_id",
          "service_id",
          "question",
          "question_revision",
          "confidence",
          "quote",
          "quote_en",
          "rationale",
          "observed_at",
          "strength",
          "decided_by",
          "status",
          "option",
          "document",
          "points",
          "feedback"
        ],
        "title": "FindingView",
        "type": "object"
      },
...
      "LeadFeedback": {
        "additionalProperties": false,
        "description": "[`LeadFeedback`](/architecture/interfaces.md#leadfeedback), the response of `API-46`.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "user_name": {
            "title": "User Name",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/LeadFeedbackVerdict"
          }
        },
        "required": [
          "id",
          "verdict",
          "note",
          "created_at",
          "user_name"
        ],
        "title": "LeadFeedback",
        "type": "object"
      },
      "LeadFeedbackVerdict": {
        "description": "`lead_feedback.verdict`.",
        "enum": [
          "RELEVANT",
          "NOT_RELEVANT",
          "ALREADY_CUSTOMER"
        ],
        "title": "LeadFeedbackVerdict",
        "type": "string"
      },
...
      "Page_AlertView_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertView"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ],
        "title": "Page[AlertView]",
        "type": "object"
      },
...
      "ScoreChange": {
        "additionalProperties": false,
        "description": "[`ScoreChange`](/architecture/interfaces.md#scorechange), one item of `API-41`.",
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccountScoreBand"
              },
              {
                "type": "null"
              }
            ]
          },
          "change_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Note"
          },
          "findings_added": {
            "items": {
              "$ref": "#/components/schemas/ScoreChangeFinding"
            },
            "title": "Findings Added",
            "type": "array"
          },
          "findings_removed": {
            "items": {
              "$ref": "#/components/schemas/ScoreChangeFinding"
            },
            "title": "Findings Removed",
            "type": "array"
          },
          "fit": {
            "title": "Fit",
            "type": "integer"
          },
          "intent": {
            "title": "Intent",
            "type": "integer"
          },
          "overrides_changed": {
            "items": {
              "$ref": "#/components/schemas/ScoreChangeOverride"
            },
            "title": "Overrides Changed",
            "type": "array"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "score_id": {
            "format": "uuid",
            "title": "Score Id",
            "type": "string"
          },
          "scoring_version": {
            "title": "Scoring Version",
            "type": "integer"
          },
          "standing": {
            "$ref": "#/components/schemas/AccountScoreStanding"
          },
          "trigger": {
            "$ref": "#/components/schemas/PipelineRunTrigger"
          }
        },
        "required": [
          "score_id",
          "as_of",
          "fit",
          "intent",
          "priority",
          "standing",
          "band",
          "scoring_version",
          "trigger",
          "run_id",
          "change_note",
          "findings_added",
          "findings_removed",
          "overrides_changed"
        ],
        "title": "ScoreChange",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LeadRadar API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/accounts/{id}/scores/{service_id}/feedback": {
      "post": {
        "description": "`API-46`: applies the api's half of [Feedback effects]\n(/architecture/rules.md#feedback-effects) to the account's lead.",
        "operationId": "post_lead_feedback_api_v1_accounts__id__scores__service_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "leadradar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadradar Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate_LeadFeedbackVerdict_"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadFeedback"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Post Lead Feedback",
        "tags": [
          "feedback-and-alerts"
        ]
      }
    },
    "/api/v1/accounts/{id}/scores/{service_id}/history": {
      "get": {
        "description": "`API-41`: the account and service's score history, newest first, each entry compared with\nthe row before it. An account with no score for the service yet answers an empty list.",
        "operationId": "get_score_history_api_v1_accounts__id__scores__service_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Service Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "leadradar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadradar Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScoreChange"
                  },
                  "title": "Response Get Score History Api V1 Accounts  Id  Scores  Service Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get Score History",
        "tags": [
          "feedback-and-alerts"
        ]
      }
    },
...
    "/api/v1/alerts": {
      "get": {
        "description": "`API-48`: newest first; `unread` true lists unacknowledged alerts only.",
        "operationId": "get_alerts_api_v1_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "service_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Id"
            }
          },
          {
            "in": "query",
            "name": "unread",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Unread"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Page Size"
            }
          },
          {
            "in": "cookie",
            "name": "leadradar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadradar Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AlertView_"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Get Alerts",
        "tags": [
          "feedback-and-alerts"
        ]
      }
    },
    "/api/v1/alerts/{id}/acknowledge": {
      "post": {
        "description": "`API-49`: acknowledging an already acknowledged alert returns it unchanged.",
        "operationId": "post_alert_acknowledge_api_v1_alerts__id__acknowledge_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "leadradar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadradar Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Post Alert Acknowledge",
        "tags": [
          "feedback-and-alerts"
        ]
      }
    },
...
    "/api/v1/findings/{id}/feedback": {
      "post": {
        "description": "`API-47`: applies the api's half of [Feedback effects]\n(/architecture/rules.md#feedback-effects) to one finding.",
        "operationId": "post_finding_feedback_api_v1_findings__id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "leadradar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Leadradar Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackCreate_FindingFeedbackVerdict_"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Post Finding Feedback",
        "tags": [
          "feedback-and-alerts"
        ]
      }
    },
...
  }
}